Removes the Action Text dependency from Active Record
Move the Action Text stuff to its own concern in `actiontext`.
