+++ actiontext/lib/action_text.rb
@@ -13,6 +13,7 @@ module ActionText
   autoload :Attachment
   autoload :Attribute
   autoload :Content
+  autoload :Encryption
   autoload :Fragment
   autoload :FixtureSet
   autoload :HtmlConversion
